First public commit
Libzt was under hobby / evening / weekend development for several
months. On January 12, 2020 I started using git for "snapshot" commits,
mainly to copy the code to other machines and operating systems to see
how it runs there. Earlier the source was just sitting on my laptop as I
was de-coupling the testing code from another hobby project.

I did my best to make libzt useful, reliable, documented and portable
but to check that I really need to open it up to users and to continue
development in the open.