Handle PVS Studio license renewal
When PVS Studio license expires, the *.PVS-Studio.log files are created
with information, that the license is stale and needs renewal.
Once the license is updated, those files stay stale, and require a
manual "make clean && make" cycle to correct.

This patch fixes that by putting modelling the dependency explicitly.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>