Add initial build support for Windows
Windows build support is pretty preliminary as I'm not accustomed to how
anything works in this environment. The build.bat script can build a
working self-test and library, assuming it is invoked from the visual
studio prompt window. The export file defines the set of public symbols,
that matches other platforms.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>