Teach the build system about various compilers
The build system had a naive expectation that each system is used
for native builds and that the default or preferred compiler of each
system is always used.

In order to support cross-compiling, or more generally, using custom
compilers. In addition some compilers support only certain features,
like being able to build a shared library but not to define the linker
export map. This patch makes the build system understand gcc, clang,
open watcom and bcc.

An build that has not used the configure script will now gain "curated"
options appropriate for each compiler.

Unless configured default to -Werror

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>