Improve support for cross-compilation
Having looked at http://crossqa.debian.net/src/libzt I realized I
totally ignored configure options related to cross compilation.

With this patch --build and --host are now parsed and stored. The main
makefile uses it to select appropriate values for CC and CXX.

I was looking at how to make this work with clang but I failed to do so,
for now. I added a safeguard to make sure we don't clobber CC and CXX if
they are set to non-default values.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>