Make makefiles modular and reusable
While working on libzc, the twin library that drives the development of
libzt, I realized that copying makefiles from libzt was not really
maintainable. As such all the makefiles powering libzt have been
converted into maintainable modules.

Some modules are just that, they can be imported and used. Other modules
describe templates that can be spawned with particular arguments.
Eventually the entire .makefiles directory can be provided externally,
through a new dependency. Right now the interface is too fragile and
untested for that.

As a major departure from the previous design, the entry point is now a
single makefile, no longer differentiated by the operating system kernel
name. The OS module provides information that is relevant to the build
system.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>