Do not test reads from short integer fields
The automatic promotion process from ZT_INTEGER to ZT_INTMAX, and from
ZT_UNSIGED to ZT_UINTMAX, was correct but the test measuring that was
flawed, as it relied on reading the shorter words instead. On some
architectures, notably Sparc 64, this results in a read of the word
zeroed by the write to the wider variant.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>